backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
lxml==4.9.3
webdriver-manager==4.0.1
requests==2.31.0

